Develop to Main (발표 이후 ~ 4월)

+++ src/main/java/com/ll/edubridge/domain/course/course/service/CourseService.java
@@ -30,10 +30,6 @@ public class CourseService {
     private final RoadmapService roadmapService;
 
 
-    public List<Course> findAll() {
-        return courseRepository.findAll();
-    }
-
     public Optional<Course> findById(Long id) {
         return courseRepository.findById(id);
     }
@@ -46,7 +42,7 @@ public class CourseService {
         Course course = Course.builder()
                 .title(createCourseDto.getTitle())
                 .notice(createCourseDto.getNotice())
-                .imgUrl(createCourseDto.getImgUrl())
+                .imgUrl("https://i.imgur.com/5JE4BGV.png")
                 .overView(createCourseDto.getOverView())
                 .price(price)
                 .writer(rq.getMember())
@@ -74,7 +70,6 @@ public class CourseService {
 
         course.setTitle(courseDto.getTitle());
         course.setNotice(courseDto.getNotice());
-        course.setImgUrl(courseDto.getImgUrl());
         course.setOverView(courseDto.getOverView());
         course.setHashtags(courseDto.getHashtags());
 
@@ -84,10 +79,16 @@ public class CourseService {
     @Transactional
     public void delete(Long id) {
         Course course = this.getCourse(id);
+
+        if(course.getConfirm()){
+            throw new GlobalException(
+                    CodeMsg.E400_12_ALREADY_HAS_ENROLL.getCode(),
+                    CodeMsg.E400_12_ALREADY_HAS_ENROLL.getMessage());
+        }
+
         courseRepository.delete(course);
     }
 
-    @Transactional
     public Course getCourse(Long id) {
         Optional<Course> course = this.findById(id);
         if (course.isPresent()) {
@@ -97,7 +98,6 @@ public class CourseService {
         }
     }
 
-    @Transactional
     public boolean haveAuthority(Long id) {
         Member member = rq.getMember();
 
@@ -107,17 +107,13 @@ public class CourseService {
 
         return false;
     }
-    public Page<Course> findByKwAdmin(KwTypeCourse kwType, String kw, Member author, Pageable pageable) {
-        return courseRepository.findByKwAdmin(kwType, kw, author, pageable);
-    }
+
     public Page<Course> findByKw(KwTypeCourse kwType, String kw, Member author, Pageable pageable) {
         return courseRepository.findByKw(kwType, kw, author, pageable);
     }
     public Page<Course> findMyCourse(Member writer, Pageable pageable){
         return courseRepository.findByWriter(writer, pageable);
     }
-
-
 
     public List<Course> findTopVotedCourses(int num){
         return courseRepository.findTopVotedCourse(num);
@@ -131,10 +127,15 @@ public class CourseService {
                     CodeMsg.E400_12_ALREADY_HAS_ENROLL.getCode(),
                     CodeMsg.E400_12_ALREADY_HAS_ENROLL.getMessage());
         }
+
+        if(!course.getConfirm()) { // 비공개 -> 공개 상황
+            String imgUrl = course.getVideoList().getFirst().getImgUrl();
+            course.setImgUrl(imgUrl);
+        }
+
         course.setConfirm(!course.getConfirm());
         return courseRepository.save(course);
     }
-
 
     public List<Course> findRecentCourse() {
         return courseRepository.findTop5ByOrderByIdDesc();
